refactor: replace AudioSystem with ModuleManager and remove system registration files
